Rename: Change style file path
